requirements.txt
-i https://pypi.org/simple
absl-py==2.2.2; python_version >= '3.8'
altair==5.5.0; python_version >= '3.9'
astunparse==1.6.3
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.0; python_version >= '3.10'
colorama==0.4.6; platform_system == 'Windows'
filelock==3.18.0; python_version >= '3.9'
flatbuffers==25.2.10
fsspec==2025.3.2; python_version >= '3.9'
gast==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
google-pasta==0.2.0
grpcio==1.71.0; python_version >= '3.9'
h5py==3.13.0; python_version >= '3.9'
huggingface-hub==0.31.2; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
keras==3.9.2; python_version >= '3.9'
libclang==18.1.1
markdown==3.8; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.5.1; python_version >= '3.9'
mpmath==1.3.0
namex==0.0.9
narwhals==1.39.1; python_version >= '3.8'
networkx==3.4.2; python_version >= '3.10'
numpy==2.1.3; python_version >= '3.10'
opt-einsum==3.4.0; python_version >= '3.8'
optree==0.15.0; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.2.1; python_version >= '3.9'
protobuf==5.29.4; python_version >= '3.8'
pyarrow==20.0.0; python_version >= '3.9'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
rich==14.0.0; python_full_version >= '3.8.0'
rpds-py==0.25.0; python_version >= '3.9'
setuptools==80.7.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.2; python_version >= '3.7'
streamlit==1.45.1; python_version >= '3.9' and python_full_version != '3.9.7'
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tensorboard==2.19.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
tensorflow==2.19.0; python_version >= '3.9'
tensorflow-io-gcs-filesystem==0.31.0; python_version < '3.12'
termcolor==3.1.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
torch==2.7.0; python_full_version >= '3.9.0'
torchvision==0.22.0; python_version >= '3.9'
tornado==6.5; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
urllib3==2.4.0; python_version >= '3.9'
watchdog==6.0.0; platform_system != 'Darwin'
werkzeug==3.1.3; python_version >= '3.9'
wheel==0.45.1; python_version >= '3.8'
wrapt==1.17.2; python_version >= '3.8'
